Fixing badmatch processing search results
If there's a discrepancy between the number of search hits
and the number of documents found for a given search to_json/3
can throw a badmatch exception on line 220 as it expects the db
and Solr to find the same number of documents. Changed the code
to assign the number of DB docs found to a new variable to
avoid the badmatch.