Add basic tests for environments/:env_name/nodes
Also had to do some cleanup surgery on a few of the test functions
that were using a function to capture setup state.  That's fine for
purely static in-memory "functional" setup, but ends up not working as
desired for side-effect-y put-something-in-the-db setup.  This is
because the setup gets evaluated all up front before actual test
execution and so other setup/cleanup actions of other tests will
interfere.  Moral of the story: use setup and foreach fixtures or be a
sadpanda.