rename pre-uninstall to uninstall, clean up help message
We are no longer doing prerm in the packages, but since we already took
the time to write code that completely stops OPC and perma-kills runit,
we might as well let users use it, e.g., for a pre-uninstall step.
Update the inline help to document the new use case.