Refactor nginx config
The external lb config is now rendered as a separate config file and
the duplication between HTTPS and HTTP lbs is removed.

A helper class NginxErb was added to make building darklaunch aware
location stanzas more concise.

The internal lb still duplicates endpoint definitions, so changes
still need to be made in two places for now.