Optimizing bucket locks
Removed calls to gproc:await/2 since these calls introduced too much
latency in the request pipeline. Now we probe the gproc registry for
the bucket name and, if not found, try to create it. If we fail to create
we try again assuming the failure was triggered by someone else creating
the bucket.