Add supervisor to start rabbitmq connection
chef_index is still not setup as an active application. The intention
for now is that an application making use of chef_index will start
this supervisor as part of its supervision tree. The starting app
specifies the name of an application that contains the rabbitmq
config. This will ease initial integration, but may not be a long term
approach as putting the config under the chef_index app might make
more sense and keeps it namespaced.