Add /_status endpoint with parallel health checking
The /_status endpoint calls the `ping/0` functions in the modules
configured in the `health_ping_modules` key of the `chef_wm`
application.

The `ping` functions should return `pong` if things are good and
`pang` if not. Each call to a `ping` health check will trigger a
timeout-induced `pang` if the call exceeds the timeout value
configured in `health_ping_timeout`.

The checks are made in parallel. The processes making the checks are
monitored by the request process to trap crashes.

Eventually, we should consider an alternative approach in which we
schedule regular health checks and report back on recent health check
history (so that a flood of /_status requets does not flood the
backend systems and because it would be easier to have all processes
properly supervised under such a setup.