Remove erl_opts from rebar.config
We want to be able to set a compiler macro in top level projects that
depend on chef_db. Currently, rebar will only honor such defines if
the subordinate project does not itself set any erl_opts in its
config. So for now, we are removing that part of the config.