Fix find_key_data to handle unset pubkey_version field
The #chef_client{} records that we are pulling from couchdb do not
have a pubkey_version set and we should assume they are certs for
now.

This is a temporary fix.

The types of key data that we need to deal with can be identified by
simple inspection of the key data itself. We should push this logic
down into chef_authn and remove the pubkey_version column and record
field.