Set sane defaults for ibrowse max sessions and pipeline
Allow 300 connections to a given server and disable request
pipelining.

We were seeing intermittent failures due to the default max session
setting in ibrowse of 10. Request pipelining in ibrowse does not
respect HTTP method and some error traces appeared to be related to
attempting to send data on a connection in the process of being
closed.