Refactor resource config and make streaming download a config option
Bookshelf config such as aws creds and whether or not to stream
downloads is now passed to the wm resource modules as init args. This
avoids application:get_env calls on each request.

Added config `stream_download` to control whether or not to use
streaming responses for GETs.

Added reset_dispatch to reload wm dispatch based on current app
config.