Call bksw_coordinator:end_read as part of bksw_io:finish_read
Since bksw_io:open_for_read initiates a read with the coordinator, we
should make update the coordinator when we're done. For most usecases,
this will happen anyway when the calling process exits, but better to
be explicit and do the right thing if the calling process is longer
lived.