Public key now returned on client create via POST
Since the URI for the client is easy to access, simply replace the
entire response body in order to return the generated key pair. This
removes the need for the append_field_to_json_body helper function and
so it has been removed.

Returning both keys makes the POST behavior align with PUT when a new
key pair is generated.