Check clients then users in fetch_requestor
We now look for a client matching the requestor name first and only
look for a matching user if no client was found. This has the
following advantages:

1. For OHC and OPC, it prevents unrelated users from masking clients
   in an org.

2. In all cases, we expect a far larger proportion of requests to come
   from clients, so it will be (slightly) more efficient to check
   clients first.

An upgrade to the signing protocol will be coming soon that will
provide the type of requestor (user or client) as part of the request
which will further cleanup this code.