Add compiler defines for oc_chef_wm to rebar.config
We need these defines at the top-level project because when we build
with locked deps, the compilation order changes and we don't inherit
oc_chef_wm's rebar.config settings when building chef_wm.

It's unfortunate that we need to duplication the settings, but moving
them out of oc_chef_wm would make local build fail.