Use chef_json as a wrapper for JSON encode/decode
This allows us to provide a normalized error message when we encounter
bad JSON upon decode. The jiffy library gives {error, Term} with
various values of Term that makes matching for invalid_json difficult
to do in a genric fashion. Using the wrapper also gives us future
flexibility if we decide to use a different JSON parsing library that
has different error semantics.