Fix parser generation and use a tag of neotoma
The latest neotoma tag includes the ability to generate a neotoma
escript that can be used to generate parsers from a grammar file. When
provided a transform module, the name of the transform module is
embedded in the generated parser. The transform module is heavily
customized and should not be deleted prior to parser
generation. Neotoma will generate a stub for you if this file is
missing, but that is not what we want.

One thing to improve is that the generated parsers will not have a
license header. We could add some additional scripting to put that
back in place after generation, but that is not included here.