Make itest depend on 'test' target, not 'compile'
Because some of the itests rely on being compiled with the TEST macro set, and rebar only does this when compiling eunit tests, we'll make the itest target depend on the 'test' target instead.  This also means we need to add '.eunit' to the path, instead of 'ebin'.

This is not an ideal solution, but it has the advantage of being easy to implement.  It also does not "pollute" the ebin with TEST-compiled BEAM files.

Our entire integration testing situation needs to be overhauled, but this will work for the time being.