Client/User should accept null public key as if undefined
Some clients will send { public_key: null } and expect the API to
preserve the public key in the datastore.

Because we are using the fun_match valid_public_key, it will not
accept a null value. The kludge is to delete any null public_key before
it gets validated by ej:valid().

This should be removed when ej:valid() fun_match (or equivalent) allows
null