Filter cookbooks using env constraints prior to dep solving
Fixes CHEF-3699

The second arg of chef_depsolver:solve_dependencies/3 is now properly
spec'd as taking [depsolver:constraint()] as returned by
chef_object:depsolver_constraints/1 and not
depsolver:dependency_set(). The implementation of
chef_depsolver:solve_dependencies/3 now pre-filters the world of
cookbooks using the environment constraints prior to calling
depsolver.

This makes the graph sent to depsolver smaller and makes the world
seen by the solver equivalent to the world shown to the user when
asking to list cookbooks available in an environment with constraints.
The downside is that the depsolver cannot provide any useful detail
about failure to solve due to constraints in the environment since it
doesn't see those versions and can't use them in backtracking.