prefer local copy of a component's software definition
Currently Omnibus does not gracefully handle a project overriding a
complete software definition. `opscode-omnibus` ships with a postgresql
software definition for version 9.1.2 but the software definition that
ships in `omnibus-software` is for version 9.2.1. Currently when Omnibus
loads the duplicate postgresql software definitions (using
instance_eval) the definitions are effectively merged and bad things
happen.

This patch is a placeholder until it can be handled in a more native way
internally in Omnibus.