Add metrics_config field to base_state record
Configuration for the top-level application name for metrics is now
retrieved from chef_wm app config. The stats_hero label function and
list of upstreams are obtained via ?BASE_RESOURCE. This isn't fully
flexibly, but it isn't clear if there would be value in exposing this
config all the way into app config. All three values are put into a
proplist and stored in the base_state record in the metrics_config
field.