Add search cache backed by redis
On startup, if app config entry use_search_cache is true, a pool of
eredis clients will be started. The number and connection details are
read from app config as:

    redis_host
    redis_port
    redis_db
    eredis_client_pool_size

A pg2 group 'redis_search_cache' is created and the clients are added
to it on start.

The oc_chef_wm_search_cache module is set as SEARCH_CACHE for use in
chef_wm.