Fix for oc-5928; jiffy:encode can return iolists as well as binaries.
bunnyc:publish (actually bunnyc:handle_call, line 169) matches Message
against type binary(). However jiffy:encode can on occasion return an
iolist.

For example the following hunk of ejson generates an iolist with jiffy.

{[{<<"silly_size">>,100000000000000000000}]}

This patch detects this and converts to binary. No test case for this
yet; I've not had a chance to do the proper mocking of bunnyc to make
this work.