Add initial specs and helpers
Lifted some things from chef-pedant.  We do not directly depend on
chef-pedant because Authz is so different from the Chef API that most
of the library methods end up either being very different or
completely unnecessary.

This commit also includes notes on the functionality that is covered
in the Cucumber tests.