Add serial IDs and use those as PK / FK targets
A bigserial is an auto-incrementing bigint, which is 8 bytes, as
opposed to 32 bytes for an authz_id.  This will keep tables smaller,
which will allow more rows to fit in the server's cache, which will
improve performance.

Also added a few helper functions that take an Authz ID and returns
the internal database ID.  These are intended for internal use.