Detect cycles by going "up" the inheritance chain
The intuition here is that going down the chain along the "child" axis
is a potentially unbounded fanout, while going up the chain along the
"parent" axis is more generally a linear chain.

It should be noted that nothing in the schema or the API currently
prevents there being a fanout of parents, but this is thought to be
the general pattern of the data.