stop `runit_service` supervise/ok race condition
Currently we wait 10 seconds for a runit service's supervise/ok named 
pipe. On slower systems (*cough* CentOS 5.x) this 10 second wait is not 
long enough.  This commit updates the embedded runit cookbook that ships 
in opscode-omnibus to match the indefinite block used in the current 
version of community cookbook:

https://github.com/opscode-cookbooks/runit/blob/1.1.0/libraries/provider_runit_service.rb#L151-L153

Longer term we need to just pull in a version of this cookbook at 
package build time using a tool like Berkshelf.