Upgrade to latest neotoma and use it as a dev-only dep
The latest version of neotoma fixes the escript build issue we had
seen in the past that had an error trace like the following:

  ```
  ==> neotoma (escriptize)
  'src/lucene_txfm.erl' already exists, skipping generation.
  escript: exception error: no match of right hand side value
  {error,enotdir}
    in function  neotoma:file/2 (src/neotoma.erl, line 42)
    in call from escript:run/2 (escript.erl, line 741)
    in call from escript:start/1 (escript.erl, line 277)
    in call from init:start_it/1
    in call from init:start_em/1
  make: *** [lucene] Error 127
  ```

The generated parsers are not updated with this commit, but verified
that they now build and the changes are minimal based on a few feature
additions to neotoma.

The Make infrastructure has been updated and re-organized and
dependencies ej and jiffy updated to latest tags used elsewhere in
erchef stack.