Add chef_index_expand to flatten/expand and POST to Solr
The module allows us to process Chef objects in the
same way as the chef-expander code does, but now we can POST to Solr
directly.

The module also enables a bulk POST mode (not yet tested against Solr)
that should allow us to index or delete multiple objects using a
single HTTP request. For migration or re-indexing purposes this should
be a big improvement as we could easily reduce the number of HTTP
calls by 10x.