Add an API to delete the search index by organization and object type.
The migration utility needs to be able to clear the index as part of
reverting a failed migration. However, nodes aren't being migrated, so
we need to be more selective about our deletions.

This adds the api delete_search_db_by_type which takes a binary
specifying the type to delete; e.g. client, data_bag_item,
environment, node or role.