Don't use lager parse transform globally
The rebar.config.lock file ends up placing lager after emysql in the
deps list, which ends up crashing the build because the lager parse
transform can't be found (because it hasn't been compiled yet).

To fix this, we remove the global compile flag for the lager parse
transform, placing it instead directly in the bifrost_request_logger
module, which is where it is actually needed.