Make search cache more robust to redis client errors
Run the eredis client queries in try/catch so that gen_server:call
timeout can be captured and returned as an error tuple.

Reduce the timeout for talking to redis to 200 ms. This is based on
measurements of how long the redis calls take and observation of
perc90 for cummulative db time for search requests. While a slow cache
still removes load from the underlying db, it doesn't make sense to
tie up failing cache requests for 5000 ms.