Add with_new_http_client/1 to oc_chef_authz_http
* Provide the plumbing needed to give callers a means to issue any
number of requests on a newly established HTTP connection to the
upstream authz service which will be cleaned up upon completion of the
callers actions.

* Refactor oc_chef_authz config format for the authz service. Bundle
authz service root URL and call timeout into a proplist as:

    {authz_service, [{root_url, URL}, {timeout, Timeout}]}

* Update the ping function to use the new configuration and take
advantage of with_new_http_client so that a new HTTP connection is
used for each ping request.

* Standardize on binary() format for ibrowse respone body return type.

* Enhance logging for failed ping requests.

* Pass req_id as header to upstream in oc_chef_authz_http if present
The req_id is an identifier for the request chain. If a value other
than 'undefined' is provided, it will be passed along to the upstream
authz service in HTTP header X-Request-Id.