WIP: open account DETs tables once at start
I think concurrent openers for DETs tables is fine. However, because
we don't open them read-only, a previous crash or unclean exit of some
kind can leave the dets file in need of repair which happens
automatically on next open. I'm unsure of how long that takes and it
seems possible that repair may not behave well under concurrent
openers?