Tune default Solr `merge_factor` down to 25.
The original `merge_factor` of 100 that seems to be the default in OPC
and OSC was cargo-culted from OHC where it was tuned to an obscenely
high value in order to optimize for sheer indexing throughput at all
costs. This eventually blew out Solr RAM in a way that the GC could not
be tuned around and simply had too many objects in the tenured
generation to be able to hold it all on the server. We tuned the
`merge_factor` down to 25 and that relieved the memory pressure
substantially, the tradeoff was that merges would happen a few times a
day, rather than once a week or so.