Reindex from a hidden node.
Not super-essential now, since Erchef isn't a distributed application,
but who knows what the future might bring?  This seems like a good bit
of hygiene for maintenance / support script, though, which really
shouldn't be considered an active participant in the node / cluster to
which it is connecting.

Erlang campfire ghost story: "THE CALL WAS COMING FROM INSIDE THE
CLUSTER!" :)