Use new LWRPs to clean up the PostgreSQL recipe.
This refactoring makes use of the new pg_user, pg_database,
pg_cluster, and pg_upgrade LWRPs.  Additionally, it separates Erchef-
and Bifrost-specific database setup tasks into their own recipes,
which are now included at the bottom.

The pg_upgrade LWRP is added here to enable cluster upgrades when
necessary.  This is the only new functionality added in this commit.

The contents of postgresql.rb are now chiefly concerned with "global"
properties of the RDBMS as a whole.