Add a pg_upgrade LWRP
This encapsulates all the logic of whether or not we'll need to run
pg_upgrade on a user's database cluster.

It determines if there is a new PostgreSQL data directory defined by
comparing with the value set in the last Chef run.  If a new one is
detected, a cluster is initialized and pg_upgrade is run with all the
appropriate flags.  The nice thing about this is that it's all
idempotent.