add depsolver as a direct dependency
The depsolver library was not configured as a
direct dependency of chef_db even though it was
utilized for filtering cookbook versions. This
caused failures while deploying chef_ojbects with
the ruby-based depsolver since we removed
depsolver entirely from the dependency tree. Since
chef_db only uses the filtering functionality of
depsolver, we can safely add it in here as an
explicit dependency. This will not start depsolver
as an active application or use any of the solving
functionality.