Temporarily add mysql2 definition
The recent 9.0.0 release of the rake-compiler gem requires rubygems
1.8.25 or greater. This simply locks down what we were using before.

I would have added this to omnibus-software instead (and may still),
but that pulls in some different changes that appear to be
incompatible with the way we have opscode-authz set up, sadly enough.
Since opscode-authz is going away real soon anyway, this seems like
the lowest-impact way to get the master branch building cleanly again.

Once the new Enterprise Chef 11 work lands on master, we can revert
this commit.