nginx: clear darklaunch headers set by client
This will clear X-Ops-Darklaunch headers set by the client
on their way to the chef server. The chef server should
not respond to darklaunch headers set by the client. Since
we're no longer using the darklaunch sub-request to set
these headers, we'll do it the simple way here.

In the near future, the load balancer routing logic
will be re-written in lua and this block of code will
be removed. It is not expected that we will need to
darklaunch endpoints before that happens. If we do,
we should use the built-in darklaunch sub-request mechanism
that's in the nginx config.