Widen cookbook versions
There are some use-cases where versions numbers are dates formatted into
the numbers. As such, the major, minor, and patch numbers for a cookbook
versions are widened to allow for this. Changes elsewhere also accept
much larger numbers.

However, this would make the open-source erchef servers going forward
validating larger numbers, and a number of existing sites that do not
have this schema change. This schema change only reflect a first-run
install. We currently do not have an migration tooling for Open Source
Chef. For now, anyone wanting to upgrade their OSC will have to do this
by hand. This will also require dropping and recreating the three views
attached to this. See: https://gist.github.com/hosh/6441356

We will be developing open-source migration tooling so that this is no
longer necessary.