First pass at integration test framework w/ temp pg db cluster
To run:

    make itest

Assumes postgres is installed and on PATH. Forces a clean and
recompile with -DTEST so that non-exported functions can be
tested. This is of debatable value for an integration test, but since
this code needs to talk to two upstream services (pg and bifrost) it
at least allows us to do some end-to-end testing with a local pg and
not worry about having a working bifrost.

Chose to try common test instead of eunit. The temp directory
management is useful and the approach feels more appropriate for
integration type testing.