Add remsh_welcome to print user prompt for remote shells
Intended for use with a wrapper script to initiate a remote shell to a
live bookshelf node, this function can be used to print a warning
message about how to exit the remote shell without shutting down the
bookshelf node.

An alternative is to add an io:format call to a .erlang file placed in
CWD relative to script execution, but this will end up getting invoked
for any executtion of the script, not just remote shells. So an
explicit prompt function seems better.