Create a remote shell to bookshelf via bin/bookshelf attach
Attach was broken when we switched from the run_erl to erlexec method
of running the bookshelf vm to make it compatible with runit.

This patch adds the glue to start a simple remote shell to a running
bookshelf instance, pulling the cookie out of the vm.args file.

Note that this depends on a recent change in bookshelf to provide
bksw_app:remsh_welcome/0 that outputs a reminder to the user about not
using q() or init:stop() to exit the remote shell.