Use more_set_headers to set users, containers, and groups xdl header
This is temporary until the new opscode-lb code gets merged in. For now,
set this according to darklaunch settings in the private chef
attributes.

These flags are also used to image the static dark launch information,
and keeping 'couchdb_*' pattern keeps the names in the static dark
launch file matched to the dynamically generated names in the dark
launch header.