Fix update_cookbook_version_checksums_with_shared_checksums test pass
I'm unclear at what point this became broken. As far as I can tell,
the test itself was not testing a valid construct. What we're trying
to test is that in the face of multiple versions with overlapping
checksums, the checksum cleanup behavior when one of the existing
cookbook versions is updated.

For this test we have:

cbv-06
  checksums: shared-1, v6-1, v6-2
cbv07
  checksums: shared-1, v7-1, v7-2

In the test we update cbv07 and set its checksums as:

cbv07 (updated)
  checksums: v7-3, v7-2

So we expect:

  Added: v7-3
  Deleted: v7-1