Refactor itest to use common test
This is a large refactor to move to common test for running the
integration tests. Common test's directory management and reporting
with full logs is a good fit for integration testing.

With this patch, we create a new postgres "cluster" running on a new
port for each test with a private data directory. This should make it
much easier to have these tests run as part of a CI build.