Fix fetch_latest_cookbook_versions tests (they were quite broken)
First, make the tests actually run. A generator form of the assert
macro was used such that these tests were not being run.

Secondly, the helper function to generate the expected results was
returning an empty list.

Rather than mess with generating, the expected result from the specs,
simply hard-code the expected values. This makes the tests easier to
understand and keeps the implementation simple.