Fix key renaming logic
This was leaving `rkey` set to `nil` when `key` was "oc_bifrost".
What this ultimately meant was that the Bifrost superuser ID, as well
as both Bifrost database user passwords were the defaults that are set
in attributes/default.rb.  They would be the exact same insecure
values on all EC installations, absent any intervention by the user to
override them.

This is a band-aid patch; the true fix is to:

A) pick a naming convention for our attributes and stick to it.

B) Stop placing attributes into attributes/default.rb that get
generated programmatically in libraries/private_chef.rb.  Had those
dummy values not been there, we would have found out about this
instantly, because Bifrost's configuration would have been invalid.