Change signatures of delete and list to match chef_db functions
The pattern is: (object_rec(), #context{}, ...)
The rationale is that in chef_object, we will always want the object
record to be the first argument since we dispatch on it (although this
isn't strictly required). But having some sort of reason seems
helpful, so we'll go with that.

This is a backwards incompatible change and will require users of
chef_db to adjust signatures for those calls.