Fetch group authz_id from SQL based on darklaunch.
Like containers, groups really have little useful information beyond
their name and authz_id.

This is a parallel construction to the containers work, except that we
only need the authz_id. The sole reference to groups is for adding a
client to the clients group, and all we need is the authz_id of the
group to send off to authz.