Use sentinel files to indicate keepalived management
This will allow add-ons to specify HA services, too.  Basically, if
there is a 'keepalive_me' file in a service's runit directory,
keepalived will manage it.

It turns out that order doesn't matter for bringing up and shutting
down services via keepalived, which makes life easier for everyone.
At the very least, alphabetical order works.

Long-term, this is probably not the way we want to go, as sentinel
files make me sqirm; this is best thought of as a "stepping stone"
along the way to the Correct Answer.

(Note that this current approach would fail if any add-on services
have keepalived post-hooks.  This is because there currently isn't a
way for add-ons to add new attributes to the "EC node", and post-hooks
are currently driven by attributes.  However, we don't have this issue
for the foreseeable future.)