Remove search caching via Redis backed cache
The search cache was only (partially) being used in HEC. It was
initially written to take load off of CouchDB and to buy us time to
continue migration work.

Analysis of request latency did not show a huge win with using the
cache. Additionally, we never fully resolved intermittent and sporadic
crashes in the eredis client (timeouts) which resulted in 500s
returned by the API.

Finally, the caching is not tested (nor used in EC) and when run with
not 100% of the production fleet as caching nodes leads to a situation
where not all nodes are "the same".