Return error for any error in bifrost
When editing a group, oc_chef_group will attempt to update all of
actors and subgroups which are in the request.  Each of these updates
currently creates a request to bifrost, by api design, which can fail.
These failures could be due to concurrent action by another user, eg
removing rights to the group or a network error or an applicaiton error
in bifrost.  In the event of an error, oc_chef_group will continue with
the remaining requests, aggregating the error to a list of errors.
This list of errors will be logged and a error will be returned to
the consumer.  This is a best effort to apply the changes.