Dialyzer fixes and adjustments for EC
* Add EC types oc_chef_container and oc_chef_group. This is a work
around and one which I'd like to see us solve in a more elegant
fashion. It's useful to have a type spec that provides the list of
known chef-world object types. But we need a way for other projects to
extend this list and make it sort of pluggable. For now, we're
hard-coding the addition of oc_chef_container and oc_chef_group as
used in EC.

* Relax the chef_updatable_object() type spec to just require a tuple.
Forcing a match on known record types makes it impossible to add in
custom records for EC. If we can find a way to make this pluggable in
future, great. For now this will do.