Implement deleting stale authz from bifrost.
oc_chef_authz_cleanup is a gen_fsm responsible for maintaining two sets.  One
of actor authz ids and one of group authz ids.  These two sets will be drained
in batches of the size specified in the configuration oc_chef_authz,
cleanup_batch_size with intervals of length oc_chef_authz, cleanup_interval.
These two configuration items, cleanup_batch_size and cleanup_interval are
defaulted to 100 (elements per set) and 2000 (milliseconds), respectively.

The gen_fsm has two states, started and stopped.  When started, a timer is set
to go off cleanup_interval milliseconds later.  That timeout event will cause
the gen_fsm to split off cleanup_batch_size number of elements from the head of
each set, actor and group.  This results in the number of operations with the
current bifrost implementation to be 2*cleanup_batch_size.  Once those
operations complete, the timer will started again waiting for cleanup_interval
milliseconds.

If the gen_fsm is stopped, the timer is cancelled and the state is maintained.
In the stopped state, authz_ids may be added.  Furthermore, individual requests
to delete orphans can be made by calling the prune api.

The prune api will send a message to the gen_fsm to attempt to prune the next
cleanup_batch_size as discussed above.  The prune api does not alter the timer
and will execute asynchronously.  It is safe to call even if the gen_fsm is
currently processing a large number of requests.