Add init_add_items, add_item, and send_items API
The new API provides an interface to bulk item flatten/expand and POST
to solr. The interface is designed to allow either a serial
implementation or an async and parallel compute model for the
flatten/expand work (with a sync join on POST to solr).

The Solr URL is exposed as an argument to init_add_items/1 to make it
easier for us to test against multiple solr instances. In the future,
we may want to remove this argument and obtain the Solr URL from app
config.

* Cleanup command processing All POSTs to solr are now treated as
"multi" and sent inside <update>...</update>. This reduces some
special case code for dealing differentiating singleton updates from
multiple item updates.

* The output XML no longer includes newline '\n' after the
delete tag.

* Normalize error return for post_to_solr to be {error, _}

* Standardizing tagged error returns to be {error, _} makes it easier
to write consistent generic error handling code.

* Add doc and specs, rename init_add_items => init_items

* post_single and post_multi return ok instead of skip now when passed
no-op input.