Support multiple vhosts for set and delete
The previous patches allow calling code to pass a custom Solr URL to
control where data will eventually be POSTed given a single
queue (rabbitmq vhost) and single set of expanders.

This patch introduces support for multiple rabbitmq vhosts. A list of
vhosts can be provided in app config at 'rabbitmq_vhosts' (note the
's'). A bunnyc gen_server is started for each vhost registered under a
name derived from the vhost.

Calling code must now know the vhost it wishes to send to. This
provides for a different mechanism of simultaneously indexing data to
two different solr instances. Here, data is sent by the client to two
vhosts and there are then assumed to be two separate sets of expanders
each talking to a different Solr instance.