Redo darklaunch switching for testing solr via multiple vhosts
Remove toggle for disabling default vhost. For now, we will always
send data to the default rabbitmq vhost.

Introduce darklaunch key "rabbit_aux_vhost". If true, data for
indexing is sent to the auxillary rabbitmq vhost for indexing.

This relies on a new app config value to map default and aux to vhost
names:

    {rabbitmq_vhosts, [{default, <<"/chef">>},
                       {aux, <<"/chef-solr4">>}]}