Create bookshelf data dir and setup ownership before running migration
To ensure correct permissions of bookshelf data without forcing a
'chown -R', we changed the user running the migration to be the
"owner" user. However, this means the migration script is unable to
create the data directory to get started.

Since the migration leaves checksum_path in place and the recipe now
creates data_path, we now guard based on the presence of the bookshelf
disk format file '_%_BOOKSHELF_DISK_FORMAT'.