Use compile mode and avoid /usr/bin/env for cookbook migration escript
Using `-mode(compile)` in an escript causes the code to be compiled
prior to execution and provides for more informative error messages.

Hard-coding the path to the escript binary simplifies the logic. The
reader no longer needs to track down that /opt/opscode/embedded/bin is
added to PATH in the default recipe. Also, this way we do not rely on
/usr/bin/env.