Fix bookshelf directory creation problems in HA configurations.
We previously created the data_path directory, ran cookbook_migration, and then created the rest of
the directories. This hit a snag because in a non-HA config we would create the data_path directory
as something like "/var/opt/opscode/bookshelf/data" and as a byproduct create the
"/var/opt/opscode/bookshelf" directory. The cookbook_migration script would see this dir, and be happy.

In HA configs the data_path directory was instead "/var/opt/opscode/drbd/data/bookshelf", which
meant that we never created "/var/opt/opscode/bookshelf", and the cookbook_migration script would
break.

We probably will need to make sure that this change doesn't impact the cookbook_migration process; I
believe that it will be fine, but I've not tested it on a system with data.

This also points up the dire need to automate our HA setup process; it is highly manual, and so CI
can't test it.