Fix 'private-chef-ctl password' ARGV validation
When private-chef-ctl was modified to use the omnibus-ctl code (since
EC 1.4.6), the password command was not properly updated.

Original fix proposed in:

    https://github.com/opscode/opscode-omnibus/pull/121

This patch is Andrew's patch minus the addition of password length
validation. Password policy should be configurable and enforced within
the API so it is consistent across tools.