Rename redis to redis_lb
This renames the redis service to redis_lb; it's easier to start with
a distinct name as the redis service for LB is substantially different
in usage than other possible users, and changing the name later as a
migration is not pleasant.

NOTE: one gotcha. In
files/private-chef-cookbooks/private-chef/libraries/private_chef.rb
We have a fixup stage for naming convention that renames '_' to '-' We
needed to override that to use redis_lb for a name.