lb: Ensure that we do not attempt to match a nil chef client version
In EHC, if chef client version is missing we do not route the request
to the lua script - instead we display a friendly message saying the
user probably doesn't belong here.

However because of the dual purpose webui/api hosting in EC, we don't
have that check - which means we need to add explicit handling for
missing client version.