Bring in fixes for opscode-account user/<users>/_acl endpoint, and tests
At some point in time we broke the user/<users>/_acl endpoint. This includes a fix to properly create the Ace object, and some work to handle invalid users/groups in the ACL.

This slipped through the cracks because this endpoint was not tested in pedant. Added tests to provide some basic coverage.