Remove purge
The purge implementation requires implementing another callback.  However,
there is no delivery timeline for the purge work.  It should be relatively
trivial to restore the purge work and implement the callback once it is back
on the delivery timeline.

Once purge is converted to a callback, there's no need for the worker.  This
elminates the maintenance burden going forward.