Fixes for keepalived.conf to work with 1.2.9 unicast
This uses the syntax introduced in 1.2.9 for unicast. Note that
keepalived is both limited in the accepted syntax and silently fails
back to multicast if the syntax isn't right.

A little bit of background:

Keepalived by default uses a multicast protocol to monitor
peers. However that isn't routable across networks.  Some customers
split their HA boxes across networks as part of their redundancy
strategy. We added unicast (aka point to point) connections using a
publicly proposed patch to keepalived 1.1.20. In the time between the
proposal and final merge into mainstream, the syntax got changed, and
our config file became silently broken.

This got missed by our tests for several reasons:

1) keepalived ignores bad syntax and silently uses multicast by
default. This works in our test setup because the two backend hosts
are on the same network.

2) While we do disable IPv4 networking as part of our IPv6 test setup,
we don't disable the dhclient daemons. Eventually the IPv4 address
reappears. Usually we have IPv4 back before we've managed to fully
complete the setup process, and keepalived may end up silently using
IPv4 multicast and not IPv6 multicast (or sometimes a mix of both) So
it is not sufficient to check that no multicast IPv6 packets are being
sent.

The only reliable way to verify that we are indeed doing the right
thing is to tcpdump the output. Unfortunately the tcpdump versions on
centos 6.2 and other older linux versions doesn't allow direct
filtering for protocol 112.

There should be no packets going to 224.0.0.18, protocol 112, nor
FF02:0:0:0:0:0:0:12, protocol 112.

http://tools.ietf.org/search/rfc3768